negative bitboard bug fix

+++ functions/bitboard.py
@@ -1,9 +1,8 @@
 import numpy as np
 
 def getBitBoard(position):
-    bitboard = np.int64(0)
-
+    bitboard = 0
     for i in position:
-        bitboard |= (np.int64(1) << (i))
-    
+        bitboard |= (1 << (i))
+
     return bitboard
